Revert "Extract CommandLineClient into abstract class"
This reverts commit ec0da39d7413c9dd88de5df699ab3452424118a9.
