--- java/clients/src/main/java/sleeper/clients/CompactionJobStatusCommandLineClient.java
@@ -1,31 +1,0 @@
-/*
- * Copyright 2022 Crown Copyright
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package sleeper.clients;
-
-import com.amazonaws.services.s3.AmazonS3;
-import sleeper.configuration.properties.InstanceProperties;
-import sleeper.configuration.properties.table.TableProperties;
-
-public abstract class CompactionJobStatusCommandLineClient extends AbstractCommandLineClient {
-
-    protected CompactionJobStatusCommandLineClient(AmazonS3 s3Client, InstanceProperties instanceProperties) {
-        super(s3Client, instanceProperties);
-    }
-
-    protected void runQueries(TableProperties tableProperties) {
-    }
-}
