+++ java/clients/src/main/java/sleeper/clients/QueryCommandLineClient.java
@@ -19,6 +19,7 @@ import com.amazonaws.services.s3.AmazonS3;
 import org.apache.commons.codec.binary.Base64;
 import sleeper.configuration.properties.InstanceProperties;
 import sleeper.configuration.properties.table.TableProperties;
+import sleeper.configuration.properties.table.TablePropertiesProvider;
 import sleeper.core.key.Key;
 import sleeper.core.range.Range;
 import sleeper.core.range.Range.RangeFactory;
@@ -31,6 +32,8 @@ import sleeper.core.schema.type.LongType;
 import sleeper.core.schema.type.PrimitiveType;
 import sleeper.core.schema.type.StringType;
 import sleeper.query.model.Query;
+import sleeper.statestore.StateStoreException;
+import sleeper.table.job.TableLister;
 
 import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
@@ -43,10 +46,34 @@ import static sleeper.configuration.properties.table.TableProperty.TABLE_NAME;
 /**
  * Allows a user to enter a query from the command line.
  */
-public abstract class QueryCommandLineClient extends AbstractCommandLineClient {
+public abstract class QueryCommandLineClient {
+    private final AmazonS3 s3Client;
+    private final TablePropertiesProvider tablePropertiesProvider;
+    private final InstanceProperties instanceProperties;
 
     protected QueryCommandLineClient(AmazonS3 s3Client, InstanceProperties instanceProperties) {
-        super(s3Client, instanceProperties);
+        this.s3Client = s3Client;
+        this.instanceProperties = instanceProperties;
+        this.tablePropertiesProvider = new TablePropertiesProvider(s3Client, instanceProperties);
+    }
+
+    public void run() throws StateStoreException {
+        TableProperties tableProperties = getTableProperties();
+        init(tableProperties);
+
+        runQueries(tableProperties);
+    }
+
+    protected abstract void init(TableProperties tableProperties) throws StateStoreException;
+
+    protected abstract void submitQuery(TableProperties tableProperties, Query query);
+
+    protected TableProperties getTableProperties() {
+        String tableName = getTableName(s3Client, instanceProperties);
+        if (tableName == null) {
+            return null;
+        }
+        return tablePropertiesProvider.getTableProperties(tableName);
     }
 
     protected void runQueries(TableProperties tableProperties) {
@@ -184,6 +211,40 @@ public abstract class QueryCommandLineClient extends AbstractCommandLineClient {
         }
         Region region = new Region(ranges);
         return new Query.Builder(tableName, UUID.randomUUID().toString(), region).build();
+    }
+
+    private String getTableName(AmazonS3 s3Client, InstanceProperties instanceProperties) {
+        List<String> tables = new TableLister(s3Client, instanceProperties).listTables();
+        String tableName;
+        if (tables.isEmpty()) {
+            System.out.println("There are no tables. Please create one and add data before running this class.");
+            return null;
+        }
+        if (tables.size() == 1) {
+            tableName = tables.get(0);
+            System.out.println("Querying table " + tableName);
+        } else {
+            Scanner scanner = new Scanner(System.in, StandardCharsets.UTF_8.displayName());
+            while (true) {
+                System.out.println("The system contains the following tables:");
+                tables.forEach(System.out::println);
+                System.out.println("Which table do you wish to query?");
+                tableName = scanner.nextLine();
+                if (tables.contains(tableName)) {
+                    break;
+                } else {
+                    System.out.println("Invalid table, try again");
+                }
+            }
+        }
+
+        System.out.println("Thie table has schema " + tablePropertiesProvider.getTableProperties(tableName).getSchema());
+
+        return tableName;
+    }
+
+    protected InstanceProperties getInstanceProperties() {
+        return instanceProperties;
     }
 
     public Key deserialise(List<String> rowKeys, Schema schema) {
